Integrate bybit linear in the rest library. Bybit add indexes and premium index.
