+++ rest/api2.go
@@ -1,41 +1,8 @@
 package rest
 
 import (
-	sjson "encoding/json"
 	"net/http"
-	"time"
 )
-
-type FundingResult struct {
-	RetCode int     `json:"ret_code"`
-	RetMsg  string  `json:"ret_msg"`
-	ExtCode string  `json:"ext_code"`
-	ExtInfo string  `json:"ext_info"`
-	Result  Funding `json:"result"`
-	TimeNow string  `json:"time_now"`
-}
-
-type Funding struct {
-	CurrentPage  int           `json:"current_page"`
-	Data         []FundingData `json:"data"`
-	FirstPageUrl string        `json:"first_page_url"`
-	From         int           `json:"from"`
-	LastPage     int           `json:"last_page"`
-	LastPageUrl  string        `json:"last_page_url"`
-	NextPageUrl  string        `json:"next_page_url"`
-	Path         string        `json:"path"`
-	PerPage      sjson.Number  `json:"per_page"`
-	PrevPageUrl  string        `json:"prev_page_url"`
-	To           int           `json:"to"`
-	Total        int           `json:"total"`
-}
-
-type FundingData struct {
-	Id     int          `json:"id"`
-	Symbol string       `json:"symbol"`
-	Value  sjson.Number `json:"value"`
-	Time   time.Time    `json:"time"`
-}
 
 // To use this functions will need to set b.BaseUrl to api2.bybit.com or api2-testnet.bybit.com
 
@@ -58,8 +25,8 @@ func (b *ByBit) GetFunding(symbol string, page int, limit int) (query string, re
 
 // GetPriceIndex
 // https://api2.bybit.com/api/price/index?symbol=BTCUSD&resolution=1&from=1605087277&to=1605173738
-func (b *ByBit) GetPriceIndex(symbol string, resolution int, from string, to string) (query string, resp []byte, result []FundingData, e error) {
-	var ret FundingResult
+func (b *ByBit) GetPriceIndex(symbol string, resolution int, from string, to string) (query string, resp []byte, result []IndexOHLC, e error) {
+	var ret IndexOHLCResult
 	params := map[string]interface{}{}
 	params["symbol"] = symbol
 	params["resolution"] = resolution
@@ -69,14 +36,14 @@ func (b *ByBit) GetPriceIndex(symbol string, resolution int, from string, to str
 	if e != nil {
 		return
 	}
-	result = ret.Result.Data
+	result = ret.Result
 	return
 }
 
 // GetPremiumIndex
 // https://api2.bybit.com/api/premium-index-price/index?symbol=BTCUSD&from=1605087277&resolution=1&to=1605173738
-func (b *ByBit) GetPremiumIndex(symbol string, resolution int, from string, to string) (query string, resp []byte, result []FundingData, e error) {
-	var ret FundingResult
+func (b *ByBit) GetPremiumIndex(symbol string, resolution int, from string, to string) (query string, resp []byte, result []IndexOHLC, e error) {
+	var ret IndexOHLCResult
 	params := map[string]interface{}{}
 	params["symbol"] = symbol
 	params["resolution"] = resolution
@@ -86,7 +53,7 @@ func (b *ByBit) GetPremiumIndex(symbol string, resolution int, from string, to s
 	if e != nil {
 		return
 	}
-	result = ret.Result.Data
+	result = ret.Result
 	return
 }
 
@@ -109,8 +76,8 @@ func (b *ByBit) LinearGetFunding(symbol string, page int, limit int) (query stri
 
 // LinearGetPriceIndex
 // https://api2.bybit.com/api/linear/public/kline/price?symbol=BTCUSDT&from=1607360460&to=1610006520&resolution=30
-func (b *ByBit) LinearGetPriceIndex(symbol string, resolution int, from string, to string) (query string, resp []byte, result []FundingData, e error) {
-	var ret FundingResult
+func (b *ByBit) LinearGetPriceIndex(symbol string, resolution int, from string, to string) (query string, resp []byte, result []IndexOHLC, e error) {
+	var ret IndexOHLCResult
 	params := map[string]interface{}{}
 	params["symbol"] = symbol
 	params["resolution"] = resolution
@@ -120,14 +87,14 @@ func (b *ByBit) LinearGetPriceIndex(symbol string, resolution int, from string, 
 	if e != nil {
 		return
 	}
-	result = ret.Result.Data
+	result = ret.Result
 	return
 }
 
 // LinearGetPremiumIndex
 // https://api2.bybit.com/api/linear/public/kline/premium-price?symbol=BTCUSDT&from=1607364960&to=1610011020&resolution=30
-func (b *ByBit) LinearGetPremiumIndex(symbol string, resolution int, from string, to string) (query string, resp []byte, result []FundingData, e error) {
-	var ret FundingResult
+func (b *ByBit) LinearGetPremiumIndex(symbol string, resolution int, from string, to string) (query string, resp []byte, result []IndexOHLC, e error) {
+	var ret IndexOHLCResult
 	params := map[string]interface{}{}
 	params["symbol"] = symbol
 	params["resolution"] = resolution
@@ -137,6 +104,6 @@ func (b *ByBit) LinearGetPremiumIndex(symbol string, resolution int, from string
 	if e != nil {
 		return
 	}
-	result = ret.Result.Data
+	result = ret.Result
 	return
 }
